implemented delta
i added my delta riddle to the mix.  for the moment it just reads from
the delta text file and prints it out.  i also added a line to check if
the user entered the correct answer for the riddle, tho i didn't know
how much to do with it so it isn't much.